+++ packages/moonstone/Scroller/Scroller.js
@@ -28,7 +28,46 @@ import useChildPropsDecorator from '../Scrollable/useChildPropsDecorator';
 import Skinnable from '../Skinnable';
 
 
-import {ScrollerBase, useSpottableScroller} from './useSpottableScroller';
+import {useSpottableScroller} from './useSpottableScroller';
+
+const ScrollerBase = {};
+
+ScrollerBase.displayName = 'ScrollerBase';
+
+ScrollerBase.propTypes = /** @lends moonstone/Scroller.ScrollerBase.prototype */ {
+	/**
+	 * Called when [Scroller]{@link moonstone/Scroller.Scroller} updates.
+	 *
+	 * @type {function}
+	 * @private
+	 */
+	onUpdate: PropTypes.func,
+
+	/**
+	 * `true` if rtl, `false` if ltr.
+	 *
+	 * @type {Boolean}
+	 * @private
+	 */
+	rtl: PropTypes.bool,
+
+	/**
+	 * Called when [Scroller]{@link moonstone/Scroller.Scroller} should be scrolled
+	 * and the focus should be moved to a scrollbar button.
+	 *
+	 * @type {function}
+	 * @private
+	 */
+	scrollAndFocusScrollbarButton: PropTypes.func,
+
+	/**
+	 * The spotlight id for the component.
+	 *
+	 * @type {String}
+	 * @private
+	 */
+	spotlightId: PropTypes.string
+};
 
 /**
  * A Moonstone-styled Scroller, Scrollable applied.
