--- packages/ui/Scroller/ScrollerBase.js
@@ -1,257 +1,0 @@
-import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import React, {} from 'react';
-
-import css from './Scroller.module.less';
-
-/**
- * An unstyled base scroller component.
- *
- * In most circumstances, you will want to use the Scrollable version.
- *
- * @class ScrollerBase
- * @memberof ui/Scroller
- * @ui
- * @public
- */
-class ScrollerBase extends React.Component {
-	static displayName = 'ui:ScrollerBase'
-
-	static propTypes = /** @lends ui/Scroller.ScrollerBase.prototype */ {
-		children: PropTypes.node.isRequired,
-
-		/**
-		 * Callback method of scrollTo.
-		 * Normally, `Scrollable` should set this value.
-		 *
-		 * @type {Function}
-		 * @private
-		 */
-		cbScrollTo: PropTypes.func,
-
-		/**
-		 * Direction of the scroller.
-		 *
-		 * Valid values are:
-		 * * `'both'`,
-		 * * `'horizontal'`, and
-		 * * `'vertical'`.
-		 *
-		 * @type {String}
-		 * @default 'both'
-		 * @public
-		 */
-		direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
-
-		/**
-		 * Prop to check context value if Scrollbar exists or not.
-		 *
-		 * @type {Boolean}
-		 * @private
-		 */
-		isVerticalScrollbarVisible: PropTypes.bool,
-
-		/**
-		 * `true` if RTL, `false` if LTR.
-		 *
-		 * @type {Boolean}
-		 * @private
-		 */
-		rtl: PropTypes.bool
-	}
-
-	static defaultProps = {
-		direction: 'both'
-	}
-
-	constructor (props) {
-		super(props);
-
-		props.setUiChildAdapter(this);
-	}
-
-	componentDidMount () {
-		this.calculateMetrics();
-	}
-
-	componentDidUpdate (prevProps) {
-		this.calculateMetrics();
-		if (this.props.isVerticalScrollbarVisible && !prevProps.isVerticalScrollbarVisible) {
-			this.forceUpdate();
-		}
-	}
-
-	scrollBounds = {
-		clientWidth: 0,
-		clientHeight: 0,
-		scrollWidth: 0,
-		scrollHeight: 0,
-		maxLeft: 0,
-		maxTop: 0
-	}
-
-	scrollPos = {
-		top: 0,
-		left: 0
-	}
-
-	getScrollBounds = () => this.scrollBounds
-
-	getRtlPositionX = (x) => {
-		if (this.props.rtl) {
-			return (platform.ios || platform.safari) ? -x : this.scrollBounds.maxLeft - x;
-		}
-		return x;
-	}
-
-	// for Scrollable
-	setScrollPosition (x, y) {
-		const node = this.props.uiChildContainerRef.current;
-
-		if (this.isVertical()) {
-			node.scrollTop = y;
-			this.scrollPos.top = y;
-		}
-		if (this.isHorizontal()) {
-			node.scrollLeft = this.getRtlPositionX(x);
-			this.scrollPos.left = x;
-		}
-	}
-
-	// for ScrollableNative
-	scrollToPosition (x, y) {
-		this.props.uiChildContainerRef.current.scrollTo(this.getRtlPositionX(x), y);
-	}
-
-	// for ScrollableNative
-	didScroll (x, y) {
-		this.scrollPos.left = x;
-		this.scrollPos.top = y;
-	}
-
-	getNodePosition = (node) => {
-		const
-			{left: nodeLeft, top: nodeTop, height: nodeHeight, width: nodeWidth} = node.getBoundingClientRect(),
-			{left: containerLeft, top: containerTop} = this.props.uiChildContainerRef.current.getBoundingClientRect(),
-			{scrollLeft, scrollTop} = this.props.uiChildContainerRef.current,
-			left = this.isHorizontal() ? (scrollLeft + nodeLeft - containerLeft) : null,
-			top = this.isVertical() ? (scrollTop + nodeTop - containerTop) : null;
-
-		return {
-			left,
-			top,
-			width: nodeWidth,
-			height: nodeHeight
-		};
-	}
-
-	isVertical = () => {
-		return (this.props.direction !== 'horizontal');
-	}
-
-	isHorizontal = () => {
-		return (this.props.direction !== 'vertical');
-	}
-
-	calculateMetrics () {
-		const
-			{scrollBounds} = this,
-			{scrollWidth, scrollHeight, clientWidth, clientHeight} = this.props.uiChildContainerRef.current;
-		scrollBounds.scrollWidth = scrollWidth;
-		scrollBounds.scrollHeight = scrollHeight;
-		scrollBounds.clientWidth = clientWidth;
-		scrollBounds.clientHeight = clientHeight;
-		scrollBounds.maxLeft = Math.max(0, scrollWidth - clientWidth);
-		scrollBounds.maxTop = Math.max(0, scrollHeight - clientHeight);
-	}
-
-	render () {
-		const
-			{className, style, ...rest} = this.props,
-			mergedStyle = Object.assign({}, style, {
-				overflowX: this.isHorizontal() ? 'auto' : 'hidden',
-				overflowY: this.isVertical() ? 'auto' : 'hidden'
-			});
-
-		delete rest.cbScrollTo;
-		delete rest.direction;
-		delete rest.rtl;
-		delete rest.isHorizontalScrollbarVisible;
-		delete rest.isVerticalScrollbarVisible;
-
-		return (
-			<div
-				{...rest}
-				className={classNames(className, css.scroller)}
-				ref={this.props.uiChildContainerRef}
-				style={mergedStyle}
-			/>
-		);
-	}
-}
-
-ScrollerBase.displayName = 'ui:ScrollerBase';
-
-ScrollerBase.propTypes = /** @lends ui/Scroller.ScrollerBase.prototype */ {
-	children: PropTypes.node.isRequired,
-
-	/**
-	 * Callback method of scrollTo.
-	 * Normally, `Scrollable` should set this value.
-	 *
-	 * @type {Function}
-	 * @private
-	 */
-	cbScrollTo: PropTypes.func,
-
-	/**
-	 * Direction of the scroller.
-	 *
-	 * Valid values are:
-	 * * `'both'`,
-	 * * `'horizontal'`, and
-	 * * `'vertical'`.
-	 *
-	 * @type {String}
-	 * @default 'both'
-	 * @public
-	 */
-	direction: PropTypes.oneOf(['both', 'horizontal', 'vertical']),
-
-	/**
-	 * Prop to check context value if Scrollbar exists or not.
-	 *
-	 * @type {Boolean}
-	 * @private
-	 */
-	isHorizontalScrollbarVisible: PropTypes.bool,
-
-	/**
-	 * Prop to check context value if Scrollbar exists or not.
-	 *
-	 * @type {Boolean}
-	 * @private
-	 */
-	isVerticalScrollbarVisible: PropTypes.bool,
-
-	/**
-	 * `true` if RTL, `false` if LTR.
-	 *
-	 * @type {Boolean}
-	 * @private
-	 */
-	rtl: PropTypes.bool,
-
-	setUiChildAdapter: PropTypes.func,
-
-	uiChildContainerRef: PropTypes.object
-};
-
-ScrollerBase.defaultProps = {
-	direction: 'both'
-};
-
-export default ScrollerBase;
-export {
-	ScrollerBase
-};
